Got tests "working" using truffle develop accounts.
I wrote a small function that signs messages the same way that a
Trezor appears to, but it needs access to private keys to do so.
Luckily, truffle's development network publishes its private keys for
its startup accounts. The tests in 'testTruffleTrezorMultiSig2o3.js'
use this function to generate signing keys in lieu of a hardware
wallet for unit testing. While this does not replace the need to
actually test the system with a real Trezor as part of the release
cycle, it does alow us to get into the nooks and cranies of the
contract with automated tests.

To run truffle tests:

```bash
$ npm install
$ truffle develop
truffle> migrate --reset
truffle> test file/testTruffleTrezorMultiSig2of3.js
...
```

Unfortunately, truffle's clean-room test isolation does not seem to be
enough to keep the tests from interacting with each other. While the
tests will all succeed individually, occasionally when they are all
run together as a set, some will fail at random because of interaction
issues.