V2.0 (#4)
* Got tests "working" using truffle develop accounts.

I wrote a small function that signs messages the same way that a
Trezor appears to, but it needs access to private keys to do so.
Luckily, truffle's development network publishes its private keys for
its startup accounts. The tests in 'testTruffleTrezorMultiSig2o3.js'
use this function to generate signing keys in lieu of a hardware
wallet for unit testing. While this does not replace the need to
actually test the system with a real Trezor as part of the release
cycle, it does alow us to get into the nooks and cranies of the
contract with automated tests.

To run truffle tests:

```bash
$ npm install
$ truffle develop
truffle> migrate --reset
truffle> test file/testTruffleTrezorMultiSig2of3.js
...
```

Unfortunately, truffle's clean-room test isolation does not seem to be
enough to keep the tests from interacting with each other. While the
tests will all succeed individually, occasionally when they are all
run together as a set, some will fail at random because of interaction
issues.

* Added wallet type selector and detect it on address form submit

* ledger address pulling script working

* ledger signing script, update to contract with tests

* wallet type selection inputs for spend() with tests

* can create a contract by pulling ledger addresses, signing partially working

* signature display working, full sig displayed that can be re-parsed

* spending with ledger successful

* manual signature entry working

* update test signatures to be the same as ledger-restricted paths

* update migration contract, update truffle version, remove hash creation
from makefile

* update to ganache, cleanly compiles and tests, remove unused packages

* removing trezor-specific signing, all tests pass

* update dapp to work with new v2.0 contract

* renaming contract to remove Trezor from name

* partial text update, ledger bip32 path restriction warning

* Ledger warnings

* minimum Trezor firmware check, closes #10

* Precalulating Ledger Display is working

* update text in spending

* show suggested gas limit for spends

* update geth debugging statement

* update myth, ledger scripts, README, Makefile

* solium output clean

* purging more trezor-specific text

* bumping python package versions and fixing python scripts

* full solium clean

* update to trezor connect v5, creation working, spending not

* upgrade to trezor connect v5 complete

* navbar links open new windows

* Replace require() messages with error codes, description in comments

* updating confirmed or failed spend working, need better error copy

* spend transaction feedback and slight ledger copy update

* update events in readme and put our name in contract comments